Cache MDS matrices (#30)
This is a bit tricky since, as far as I know, Rust doesn't provide an easy way to store static data that is specific to a generic type (like `F` in our case). To work around that, I've used a single cache map with the field's TypeId in the key and "serialized" field data in the value. Let me know if you see a more elegant solution though.