Some fixes relating to our use of the Halo endo (#51)
* Some fixes relating to our use of the Halo endo

This should compute `halo_us` properly. (I'm avoiding "IPA challenges" now because of the ambiguity.)

The main method is now failing with a PI related error, but I think it makes sense. The recursive circuit expects its inner proof to publish assumptions as PIs, but the dummy proof has no assumptions or PIs.

* Fix generator per PR feedback

* Partial fix (test is still red)

* Delete unused to_vec()

* Disable failing test