Allow PIs to be added at any point in CircuitBuilder
This is a little confusing since we now have two notions of gate indices -- one for the "original" indices, before prepending `PublicInputGate`, and one for after.

Ideally, to avoide this confusion, it might be best if `Target::Wire` didn't store a gate index at all; we could instead use addresses for `Eq`, `Hash`, etc. (see `ptr::eq`, `ptr::hash`). That would be a major refactor though.